Adding flatten and restore functionality to the MCLabelContainer
Adding flatten_to and restore_from methods which allow to flatten to a container
instance and restore the vector content from a buffer, both by copy. This can
be used with a container with pmr allocator and underlying memory resource.

Flattening always requires a copy, but the data can be directly copied
to the message resource. Restore can be done using spectator memory
resources for the two regions of the input buffer, but this kind of
memory resource needs to be implemented.