Fix systematic drift bug in TRD digitization
I noticed that we were systematically loosing electrons
in the signal generation. The reason is that the locC position
is smeared by diffusion and that each electrons applied also
a linear shift.
Further, the diffusion was not independent for each electron as the locC
was merely reused in its last state.

This commit fixes this problem by making the diffusion process start always from
the original hit coordinates. The final coordinates are now
specific/local to each electrion.