Exclude orbit=0 generation and its ovefrlow by InteractionSampler
Some detectors need to inspect BC's before the one being triggered,
to exclude negative times we shall start orbit/bc sampling from orbit 1
instead of unrealistic 0 (in the real running the orbit will be reset
to 0 at the beginning of the each fill).

Produce fatal if orbit exceeds maxOrbit (>100 hours) and wraps to 0.