[EMCAL-629] Add support for in-memory raw decoding
Add RawReaderMemory (templated) for the support of
raw encoding of an in-memory block of raw data
(directly provided by the readout application).
The interface is the same as in RawReaderFile.
The raw buffer is adapted in order to construct
itself from a char array. Template specification
is added for the EMCAL raw data header.