[EMCAL-534, EMCAL-629] Make AltroDecoder templated
Make AltroDecoder templated in order to support
different versions of the raw reader and different
versions of the raw header. For this all child
classes are extracted as separate classes.
RCUTrailer, Bunch and Channel are extracted
to separate files as they are used by in the raw
encoding, the error handling of the AltroDecoder
(AltroDecoderError) is kept in the same file as
separate class. Template specifications for
RawReaderMemory and RawReaderFile with
the EMCAL raw data header are added.