Defining o2::itsmft::Cluster as messageable
In order to safely detect messageable types they must be trivially copyable.
It appears that o2::itsmft::Cluster is not trivially copyable, which is due
to the underlying ROOT classes for Point3D.

std::is_trivially_copyable<ROOT::Math::Cartesian3D> fails because the class
implements a copy constructor, although it does not much more than the default copy
constructor.

A trivially copyable class is a class (defined with class, struct or union) that:

    uses the implicitly defined copy and move constructors, copy and move assignments,
    and destructor.
    has no virtual members.
    its base class and non-static data members (if any) are themselves also trivially
    copyable types.

As one attempt, std::is_trivially_copyable has been specialized for Point3D, which is
workign correctly. However adding Poin3D as a member still makes a simple struct
non trivially copyable. Have to understand how the type trait is realized for class
members.

We believe that o2::itsmft::Cluster is messageable and explicitely specialize the
type trait. A unit test is added to go beyond make-believe