DPL: Optimization for associating output messages with actual channels
This is a follow-up to commit f6d950b1

1. Introducing a registry of channel names in the MessageContext

The unique references of string objects from the registry allow fast access
and association of messages to actual FairMQChannels by avoiding string
allocation.

2. Optimization to unordered_map

Using pointer to channel string as key to avoid allocation of strings.
The ContextObjects hold references of the string variables from the registry.
The string references from the OutputRoute definition of the device do not
fulfill the requirement to refer to one single string object per channel,
because the device stores a vector of OutputRoutes.