DPL: backoff devices with dangling inputs
Looks like we still compute dangling inputs, even when we sent an "endOfStream"
event. IIRC, the idea behind this was that we might want to resume a stream
on some external input (e.g. a new file queued to be processed). However under
no circumstances this should be considered a high frequency mode, so we always
backoff when we are in StreamingState::Idle.