Adding DPL RawParser utility
The parser handles transparently input in the format of raw pages.
There three levels:
- the DPL processor has one or more input route(s)
- multiple parts per input route (split payloads or multiple input
  specs matching the same route spec
- variable number of raw pages in one payload

The parser provides an iterator interface to loop over raw pages,
starting at the first payload at the first route and going to the
next route when all payloads are processed.

A list of InputSpec definitions can be provided to filter the incoming
messages on the DataHeader information.