Adding posibility to disable writers in tpc-reco-workflow
The running processors are configured by the type of desired output, like
e.g. raw, clusters, tracks. A writer process is by default added For each
of the configured outputs. When combining the workflow with other workflows,
the writers are not always necessary and can now be removed from the workflow
using the key 'disable-writer' in the '--output-type' option.

This can also be used to prevent a routing bug which is triggered when attaching
the DPL output proxy or the QC workflow to the TPC reco workflow.

   o2-tpc-reco-workflow ... | o2-dpl-output-proxy --dataspec "dpl-output-proxy:TPC/TRACKS ...

The proxy is added between tracker and writer, however the writer is trying to
send back data to the proxy via a channel which is not existing

[23170:tpc-track-writer]: [23:46:43][ERROR] requested channel has not been configured? check channel names/configuration.
[23170:tpc-track-writer]: [23:46:43][ERROR] channel: from_tpc-track-writer_to_dpl-output-proxy, index: 0