Create the bitfield of active sectors from the list of TPC sectors
With ZS raw data as input to the tracker, there is no TPC sector header and
the bitfield of active sectors is zero. Need to create the bitfield from the
sector numbers. We need to check if this information should be added upstream
or if this is simply not needed when the tracker reads the raw data directly.