Detector may chose if HBF end is signaled by RDH.stop on new page
Some detectors signal the end of the HBF by adding an empty CRU page containing
just a header with RDH.stop=1 while others may simply set the RDH.stop in the last
page of the HBF (it may appear to be also the 1st and the page of the HBF and may or
may not  contain the payload). This behaviour is steered by call
writer.setAddSeparateHBFStopPage(bool v). By default end of the HBF is signaled on separate page.