Make sure to print an error if data is lost while terminating DPL proxy
Amending PR #3801. Checking for the requested change of device state before
dispatching messages might silently drop data.

Have to ensure that message are always dispatched, and if the sending does not
succeed an error message is printed. Only after that the pending device state
change is checked and can break the loop.