Simple GRID submission tool
A helper script, making it easy to submit existing
local scripts as an ALIEN GRID job (with the following notation):

grid-submit my_script.sh jobname

All communication with the GRID system will be done by the script
and JDLs setup automatically.
This can also be used to emulate a GRID-like environment locally.

Mainly a convenience tool for (aliperf) testing. Not yet powerful for production
usages.